requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
kombu==5.3.7
prompt_toolkit==3.0.47
psycopg-binary==3.1.19
psycopg2==2.9.9
python-crontab==3.1.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
celery>=5.2.7
redis==5.0.0b2
